Fix race condition with widgets painting as empty
There is a small gap between reading job status, and then generating
the json representation of the job.  The job was being refreshed
to create the json, which meant the job might have changed.
Occasionally, this would cause the code to enter the "not complete"
path, but then the job actually completes before job.json()
is retrieved and the json would reflect the now completed job,
which is not correct.