Implement field_map to customize table fields
This changes the mechanism for allowing the same report to have
multiple copies of the same field.  When the table is created,
the 'field_map' parameter is specified to map fields attached to
the table to new keywords, and optionally modified field attributes.

This is used when defining data source tables that automatically
pull in standard fields that they require.  For example,
the WiresharkTable adds the field 'pcapfilename', which is extracted
from criteria by the WiresharkQuery class when the query is run.
If two tables are defined and would like to leverage different pcap
files, the field_map is used as follows:

t1 = WiresharkTable.create('table-1',
    field_map={'pcapfilename': {'keyword': 'pcapfilename_1',
                                'label': PCAP File 1'}})

t2 = WiresharkTable.create('table-2',
    field_map={'pcapfilename': {'keyword': 'pcapfilename_2',
                                'label': PCAP File 2'}})