Add FocusedAnalysisTable for use in report definitions
This analysis table will take a source table as input, will evaluate
that table for the overall max or min value, determine the time where
that occurs, and then run a secondary table with an adjusted
timeframe and resolution around that time point.

In effect, this provides a zoom capability on times of interest,
and the secondary table can be of any datasource.