Guard against check_children race condition
When the last child job marks itself complete and signals its parent
to check_children, the race condition could occur that the child
job still appears running when the parent checks children.  The
child in this case is still exiting its Transaction and thus the DB
doesn't reflect its changed status yet.

The fix adds a TransactionLock on the same child object, which
will make the parent block until that child has completed its
outer transaction leading to an accurate status.

job status that the child