Add guest read-only access
This allows reports to be run without any authentication.  The
setting can be defined in local_settings, and parameters like
the Timezone to use, and the username to display can be configured.

Most of the other operations, including help, still require an
authenticated user which is enforced on the server side.