Scheduler now uses common Connection; Fix ReportHistory job handle calcs
- Change from bare requests methods to using common.Connection
class.

- Report history was being created using the criteria set
from the POSTed report criteria.  When the actual
Jobs get created though, a subset of that criteria specific
to the table gets used instead, creating a different handle.
When that Job completed, it wouldn't match up with the RH
handles that were originally calculated.  This update
uses the same method to calculate handles as widget jobs.`